% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmpliconSequencing.R
\name{as.list.tas.object.settings}
\alias{as.list.tas.object.settings}
\title{Coerce tas.object.settings to list}
\usage{
\method{as.data.frame}{tas.dna.repair}(x)
}
\arguments{
\item{x}{A tas.object.settings object}

\item{...}{Passed to data.frame}
}
\description{
Coerce tas.object.settings to list
}
