% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchAnalyzeAmplicon.R
\name{batchAnalyzeAmplicon}
\alias{batchAnalyzeAmplicon}
\title{Batch amplicon analysis}
\usage{
batchAnalyzeAmplicon(input.settings.list, progress_callback = NULL)
}
\arguments{
\item{input.settings.list}{A character vector describing a path to a .csv file to read settings from. Alternatively, an R data.frame or list containing those settings.}

\item{progress_callback}{Optional function receiving structured progress events during analysis}
}
\value{
A list of AmpliconSequencing objects
}
\description{
Function providing for batch analysis of samples. Sample settings must be
provided, either as a .csv file (1 sample per row), or as an R object (a list
or data.frame) containing the same information. See \code{\link{readSettings}} for
more information.
}
\section{See also}{

\code{\link{batchSummarize}}
}

\examples{
\dontrun{
  batchAnalyseAmplicon('/.../folder/settings.csv')
  batchAnalyzeAmplicon(data.frame)
  batchAnalyzeAmplicon(list)
}

}
