% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmpliconSequencing.R
\name{charDisplayTrim}
\alias{charDisplayTrim}
\title{Function charDisplayTrim}
\usage{
charDisplayTrim(string)
}
\arguments{
\item{string}{character vector to be trimmed}
}
\value{
a truncated string
}
\description{
Helper function to trim longer character vectors to 12 characters and (...). Used in "show" methods for more compact display with long sequences.
}
