% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmpliconSequencing.R
\name{empty.pass}
\alias{empty.pass}
\title{Function empty.pass}
\usage{
empty.pass()
}
\value{
An empty PairwiseAlignmentsSingleSubject object.
}
\description{
Helper function to create an empty PairwiseAlignmentsSingleSubject object. Used to initialize S4 class "tas.alignment".
}
