% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmpliconSequencing.R
\name{getAIDTables}
\alias{getAIDTables}
\alias{getAIDTables,tas.aid.tables-method}
\alias{getAIDTables,AmpliconSequencing-method}
\alias{getWRCHTable}
\alias{getWRCHTable,tas.aid.tables-method}
\alias{getWRCHTable,AmpliconSequencing-method}
\alias{getWRCYTable}
\alias{getWRCYTable,tas.aid.tables-method}
\alias{getWRCYTable,AmpliconSequencing-method}
\title{Retrieve AID mutation tables}
\usage{
\S4method{getAIDTables}{tas.aid.tables}(object)

\S4method{getAIDTables}{AmpliconSequencing}(object)

\S4method{getWRCHTable}{tas.aid.tables}(object)

\S4method{getWRCHTable}{AmpliconSequencing}(object)

\S4method{getWRCYTable}{tas.aid.tables}(object)

\S4method{getWRCYTable}{AmpliconSequencing}(object)
}
\arguments{
\item{object}{An S4 object of class tas.aid.tables or AmpliconSequencing}
}
\description{
Extracts tables summarizing positions and mutations at AID hotspots from an S4 object of class tas.aid.tables or its parent class AmpliconSequencing.
}
\section{Methods (by class)}{
\itemize{
\item \code{getAIDTables(tas.aid.tables)}: Method for class tas.aid.tables

\item \code{getAIDTables(AmpliconSequencing)}: Method for class AmpliconSequencing

\item \code{getWRCHTable(tas.aid.tables)}: Method for class tas.aid.tables

\item \code{getWRCHTable(AmpliconSequencing)}: Method for class AmpliconSequencing

\item \code{getWRCYTable(tas.aid.tables)}: Method for class tas.aid.tables

\item \code{getWRCYTable(AmpliconSequencing)}: Method for class AmpliconSequencing

}}
\section{Functions}{
\itemize{
\item \code{getAIDTables()}: Extracts a list containing tables for WRCH and WRCY AID hotspot motifs

\item \code{getWRCHTable()}: Extracts a data.frame showing positions and mutations at WRCH AID hotspot motifs

\item \code{getWRCYTable()}: Extracts a data.frame showing positions and mutations at WRCY AID hotspot motifs

}}
\examples{
getAIDTables(object)
getWRCHTable(object)
getWRCYTable(object)
}
