% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics-tas.sequences.R
\name{getFilterCounts}
\alias{getFilterCounts}
\alias{getFilterCounts,tas.sequences-method}
\alias{getFilterCounts,AmpliconSequencing-method}
\title{Extrat read filter counts}
\usage{
getFilterCounts(object)

\S4method{getFilterCounts}{tas.sequences}(object)

\S4method{getFilterCounts}{AmpliconSequencing}(object)
}
\arguments{
\item{object}{An S4 object of class tas.sequences or AmpliconSequencing}
}
\value{
A data.frame containing read filter counts after each step of the analysis
}
\description{
Extrat read filter counts
}
\section{Methods (by class)}{
\itemize{
\item \code{getFilterCounts(tas.sequences)}: Method for class tas.sequences

\item \code{getFilterCounts(AmpliconSequencing)}: Method for class AmpliconSequencing

}}
\examples{
\dontrun{
  getFilterCounts(object)
}
}
