% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmpliconSequencing.R
\name{isEmpty,tas.mutations-method}
\alias{isEmpty,tas.mutations-method}
\title{isEmpty methods for S4 class tas.mutations}
\usage{
\S4method{isEmpty}{tas.mutations}(x)
}
\description{
isEmpty methods for S4 class tas.mutations
}
\section{Functions}{
\itemize{
\item \code{isEmpty(tas.mutations)}: methods for tasAnalyzer S4 class tas.mutations

}}
