% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmpliconSequencing.R
\name{isEmpty-tasAnalyzer}
\alias{isEmpty-tasAnalyzer}
\title{isEmpty methods for package tasAnalyzer}
\usage{
\S4method{isEmpty}{tas.object.settings}(x)

\S4method{isEmpty}{tas.mutations}(x)

\S4method{isEmpty}{tas.sequences}(x)

\S4method{isEmpty}{tas.aid.tables}(x)

\S4method{isEmpty}{tas.dna.repair}(x)

\S4method{isEmpty}{tas.alignment}(x)

\S4method{isEmpty}{AmpliconSequencing}(x)
}
\arguments{
\item{x}{An S4 object from tasAnalyzer}
}
\description{
Extends the S4 function isEmpty to detect whether S4 classes in tasAnalyzer are empty, holding only initialized values.
}
\section{Functions}{
\itemize{
\item \code{isEmpty(tas.object.settings)}: Method for class tas.object.settings

\item \code{isEmpty(tas.mutations)}: Methods for class tas.mutations

\item \code{isEmpty(tas.sequences)}: Methods for class tas.sequences

\item \code{isEmpty(tas.aid.tables)}: Methods for class tas.aid.tables

\item \code{isEmpty(tas.dna.repair)}: Methods for class tas.dna.repair

\item \code{isEmpty(tas.alignment)}: Methods for class tas.alignment

\item \code{isEmpty(AmpliconSequencing)}: Methods for class AmpliconSequencing

}}
