% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmpliconSequencing.R
\name{isEmpty,tas.sequences-method}
\alias{isEmpty,tas.sequences-method}
\alias{isEmpty,tas.aid.tables-method}
\alias{isEmpty,tas.dna.repair-method}
\alias{isEmpty,tas.alignment-method}
\alias{isEmpty,AmpliconSequencing-method}
\title{isEmpty function for S4 class tas.sequences}
\usage{
\S4method{isEmpty}{tas.sequences}(x)

\S4method{isEmpty}{tas.aid.tables}(x)

\S4method{isEmpty}{tas.dna.repair}(x)

\S4method{isEmpty}{tas.alignment}(x)

\S4method{isEmpty}{AmpliconSequencing}(x)
}
\description{
isEmpty function for S4 class tas.sequences

isEmpty function for S4 class tas.aid.tables

isEmpty function for S4 class tas.dna.repair

isEmpty function for S4 class tas.alignment

isEmpty function for S4 class AmpliconSequencing
}
\section{Functions}{
\itemize{
\item \code{isEmpty(tas.sequences)}: methods for tasAnalyzer S4 class tas.sequences

\item \code{isEmpty(tas.aid.tables)}: methods for tasAnalyzer S4 class tas.aid.tables

\item \code{isEmpty(tas.dna.repair)}: methods for tasAnalyzer S4 class tas.dna.repair

\item \code{isEmpty(tas.alignment)}: methods for tasAnalyzer S4 class tas.alignment

\item \code{isEmpty(AmpliconSequencing)}: methods for tasAnalyzer S4 class AmpliconSequencing

}}
