% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measureMutations.R
\name{measureMutations}
\alias{measureMutations}
\title{Measure mutations}
\usage{
measureMutations(sequence.table, settings)
}
\arguments{
\item{sequence.table}{S4 object of class tas.sequences}

\item{settings}{S4 object of class tas.settings}
}
\value{
An S4 object of class tas.mutations
}
\description{
Measure mutations
}
\examples{
\dontrun{
  measureMutations(sequence.table, settings)
}
}
