% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-PairedAmpliconSequencing.R
\docType{class}
\name{PairedAmpliconSequencing-class}
\alias{PairedAmpliconSequencing-class}
\title{S4 class PairedAmpliconSequencing}
\description{
S4 class that holds results of paired tasaR analysis. Incorporates a control and
an experimental condition for comparison.
}
\section{Slots}{

\describe{
\item{\code{Control}}{S4 object of class AmpliconSequencing}

\item{\code{Experimental}}{S4 object of class AmpliconSequencing}
}}

