% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{readSettings}
\alias{readSettings}
\title{Read settings for tasaR}
\usage{
readSettings(x, row = NULL)
}
\arguments{
\item{x}{The object to be analyzed. May be a file path ("./X.csv"), or a data.frame or list with the appropriate names and classes for the object. Function should automatically detect which type input 'x' is.}

\item{row}{If using a .csv input, the row number in the sheet that contains the information for this sample. Use the \strong{displayed number}, the function will account for and remove the header row during processing.}
}
\value{
An S4 object of class tas.object.settings
}
\description{
The tasaR package has a relatively complex set of settings that must be defined for each sample. This function reads those settings from either a .csv file on disk or an R object (list/data.frame) and generates an appropriate S4 object of class tas.object.settings.
}
\examples{
readSettings('/.../folder/settings.csv', row = 2)
readSettings(data.frame)
readSettings(list)
}
