% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-show.R
\docType{methods}
\name{show-tasaR}
\alias{show-tasaR}
\alias{show,tasaR-method}
\alias{show,tas.object.settings-method}
\alias{show,tas.mutations-method}
\alias{show,tas.sequences-method}
\alias{show,tas.dna.repair-method}
\alias{show,AmpliconSequencing-method}
\alias{show,PairedAmpliconSequencing-method}
\title{show methods for package tasaR}
\usage{
\S4method{show}{tas.object.settings}(object)

\S4method{show}{tas.mutations}(object)

\S4method{show}{tas.sequences}(object)

\S4method{show}{tas.dna.repair}(object)

\S4method{show}{AmpliconSequencing}(object)

\S4method{show}{PairedAmpliconSequencing}(object)
}
\arguments{
\item{object}{S4 object to print to console}
}
\description{
Extends the show function to print to console contents of S4 objects in the tasaR package.
}
