% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeAmplicon.R
\name{Summarize}
\alias{Summarize}
\title{Summarize amplicon results}
\usage{
Summarize(results, export = FALSE, path = NULL, suppressConsoleOutput = FALSE)
}
\arguments{
\item{results}{S4 object of class AmpliconSequencing}

\item{export}{Logical specifying whether to export results or not}

\item{path}{(Optional) Character, File path to directory to export results to}

\item{suppressConsoleOutput}{Logical, specify whether to suppress output to console if appropriate}
}
\value{
A list containing Graphs and Tables for the sequencing object
}
\description{
Convenience wrapper to graph all results from a single AmpliconSequencing
object and save them along with extracted tables to a separate list. Can
also export all results to a specified directory.
}
\examples{
\dontrun{
Summarize(AmpliconSequencing, export = TRUE, path = file.path(tempdir(), export))
}
}
