% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmpliconSequencing.R
\docType{class}
\name{tas.aid.tables-class}
\alias{tas.aid.tables-class}
\title{S4 class tas.aid.tables}
\description{
S4 object of tables summarizing mutations at AID cytosine motifs. Separate tables for the broader WRCH motif and the more specific WRCY motif.
}
\section{Slots}{

\describe{
\item{\code{WRCH}}{data.frame. Coordinates and mutations at WRCH AID cytosines.}

\item{\code{WRCY}}{data.frame. Coordinates and mutations at WRCY AID cytosines.}
}}

\section{Table Columns}{

\describe{
\item{Motif}{character vector DNA motif (top strand = WRCH/WRCY, bottom strand = DGYW/RGYW)}
\item{Start}{numeric vector of starting coordinates of AID motifs}
\item{End}{numeric vector of ending coordinates of AID motifs}
\item{Cytosine}{numeric vector of coordinates of AID cytosines}
\item{MotifMutagenesis}{numeric vector of the total frequency of mutagenesis within the motif (max 400\\%)}
\item{CytosineMutagenesis}{numeric vector of the frequency of mutagenesis at the AID cytosine}
}
}
}

