% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{tas.mutations-class}
\alias{tas.mutations-class}
\title{S4 class tas.mutations}
\description{
S4 object of sequence mutations by position
}
\section{Slots}{

\describe{
\item{\code{AllMutations}}{data.frame showing position and frequency of mutations at all positions}

\item{\code{CytosineMutations}}{data.frame showing position and frequency of mutations at AID cytosines}

\item{\code{NonCytosineMutations}}{data.frame showing position and frequency of mutations at all positions that are not AID cytosines}

\item{\code{MotifSums}}{numeric vector summarizing the frequency of mutations among different motifs and denominators}
}}

