% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmpliconSequencing.R
\docType{class}
\name{tas.object.settings-class}
\alias{tas.object.settings-class}
\title{S4 class tas.object.settings}
\description{
S4 object containing analysis settings for tasAnalyzer
}
\section{Slots}{

\describe{
\item{\code{Name}}{character vector of sequence name}

\item{\code{IsAntibody}}{logical vector if sequence is an antibody}

\item{\code{MeasureSHM}}{logical vector defining whether to measure somatic hypermutation (SHM)}

\item{\code{MeasureDNARepairTypes}}{logical vector defining whether to measure DNA repair types}

\item{\code{MergedFASTQPath}}{character vector of file path to the merged paired-end FASTQ file to be analyzed}

\item{\code{ReferenceSequence}}{character vector of the reference (expected) amplicon sequence to be analyzed (do not include primers)}

\item{\code{ForwardExtensionType}}{character vector ("barcode", "umi", or "") of forward primer extension type (if any)}

\item{\code{ForwardExtension}}{character vector of DNA sequence of the forward extension}

\item{\code{ForwardPrimer}}{character vector of DNA sequence of the forward primer (only sequence that binds to the target DNA)}

\item{\code{ReverseExtensionType}}{character vector ("barcode", "umi", or "") of reverse primer extension type (if any)}

\item{\code{ReverseExtension}}{character vector of DNA sequence of the reverse extension}

\item{\code{ReversePrimer}}{character vector of DNA sequence of the forward primer (only sequence that binds to the target DNA)}

\item{\code{AmpliconLength}}{integer vector of the total amplicon length, including primers and extensions (exclude any Illumina sequencing adapters)}

\item{\code{InsertStart}}{integer vector pointing to the coordinate of the first nt of the ReferenceSequence within the total amplicon}

\item{\code{InsertEnd}}{NEGATIVE integer vector pointing to the coordinate of the last nt of the ReferenceSequence within the total amplicon (count from the end)}

\item{\code{AntibodyRegions}}{integer vector containing coordinates of the first nt of each region of an antibody (or all NA if IsAntibody = FALSE)}
}}

