% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{tas_analyze}
\alias{tas_analyze}
\title{One-step analysis and export of targeted amplicon sequencing experiments}
\usage{
tas_analyze(
  InputFilePath,
  config = "manual",
  shiny.settings = list(),
  shiny.env = FALSE,
  shiny.fileTable = character(),
  WorkingDirectory = NULL,
  PandaseqDirectory = NULL,
  merge.reads = TRUE,
  measure.shm = FALSE,
  sequence.alignment.count = 10,
  read.frequency.limit = 0.001,
  protein.mutations = FALSE,
  PhyloTree = FALSE,
  dna.repair.pathways = FALSE,
  multicore = TRUE
)
}
\arguments{
\item{InputFilePath}{The complete file path of the input .csv file. Make sure all R1 and R2 files are in the same directory.}

\item{config}{Determines where the function draws the analysis parameters from. 'manual' allows paramaters to be specified in the function, whereas 'shiny' is used by the shiny app.}

\item{shiny.settings}{A list of function settings set by the shiny app.}

\item{shiny.env}{A logical (TRUE/FALSE) to tell whether the function is being executed by the shiny app (TRUE) or not (FALSE).}

\item{shiny.fileTable}{A character vector containing names of files uploaded to the shiny app.}

\item{WorkingDirectory}{Optional. The path to the folder containing the input .xlsx file and all input .fastq files.}

\item{PandaseqDirectory}{Optional. The path to the folder containing the pandaseq.exe files for use on Windows. Not currently implemented.}

\item{merge.reads}{A logical (TRUE/FALSE) to tell whether PANDAseq should be used to merge paired end read files.}

\item{measure.shm}{A logical (TRUE/FALSE) to tell whether the somatic hypermutation analysis module should be run.}

\item{sequence.alignment.count}{Enter a whole number value to determine how many sequences to display as multiple sequence alignments. Default value of 10 is recommended, can require significant resources for larger alignments.}

\item{read.frequency.limit}{If UMIs are not used, sequences will be filtered using this number as the minimum read frequency percentage for a sequence to be included in the analysis. Values of 0.1-0.001 are recommended.}

\item{protein.mutations}{A logical (TRUE/FALSE) to tell whether mutations in the amino acid sequences should be analyzed.}

\item{PhyloTree}{A logical (TRUE/FALSE) to tell whether phylogenetic trees should be made for the number of sequences specified by sequence.alignment.count.}

\item{dna.repair.pathways}{A logical (TRUE/FALSE) to tell whether this is a gene editing experiment and the type of mutations observed should be classified and counted by the likely underlying DNA repair pathway.}
}
\value{
A list containing tables and graphical analyses generated by the software.
}
\description{
Full analysis of a targeted amplicon sequencing experiment on Illumina, starting from R1 and R2 .fastq reads.
}
