% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{tas_export_msa}
\alias{tas_export_msa}
\title{Export multiple sequence alignment graphics}
\usage{
tas_export_msa(Sample.Names, Output.MSA.List, Config.List, WD)
}
\arguments{
\item{Sample.Names}{A character vector with names of all samples.}

\item{Output.MSA.List}{A list of multiple sequence alignment graphs for each sample. Names of each list entry must correspond to Sample.Names.}

\item{Config.List}{A list containing configuration parameters for tasAnalyzer.}

\item{WD}{A character vector indicating the file path to the working directory containing input files.}
}
\value{
No returns within R. Writes image files to disk.
}
\description{
Exports the multiple sequence alignment graphs created by tas_msa as .png (alignments) or .pdf (phylogenetic tree) objects.
}
