% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{tas_export_mutations}
\alias{tas_export_mutations}
\title{Export graphs and spreadsheets}
\usage{
tas_export_mutations(
  Sample.Names,
  Output.Mutations.List,
  Sequence.Table.List,
  Reference.Sequences.DNA,
  Input.DataFrame,
  Config.List,
  WD
)
}
\arguments{
\item{Sample.Names}{A character vector with names of all samples.}

\item{Output.Mutations.List}{A list of the analyzed mutation outputs for each sample. Names of each list entry must correspond to Sample.Names.}

\item{Sequence.Table.List}{A list of data frames containing each unique DNA sequence for all samples. Names of each list entry must correspond to Sample.Names.}

\item{Reference.Sequences.DNA}{A list containing the expected DNA sequences for each sample.}

\item{Input.DataFrame}{A data frame imported from the .xlsx file specifying details of each sample.}

\item{Config.List}{A list containing configuration parameters for tasAnalyzer.}

\item{WD}{A character vector indicating the file path to the working directory containing input files.}
}
\value{
A list adding new tables and graphs for each sample to the original Output.Mutations.List.
}
\description{
Generates graphs and workbooks showing the results of sequence labeling and mutation analysis. Writes image files and .xlsx spreadsheets to disk for archive and compatibility with other software analyses.
}
