% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{tas_filter}
\alias{tas_filter}
\title{Filtering on paired end reads.}
\usage{
tas_filter(Sample.Names, Reads.List, Input.DataFrame, Config.List)
}
\arguments{
\item{Sample.Names}{A character vector with names of all samples.}

\item{Reads.List}{A list of merged reads in ShortReads format. Names of each list entry must correspond to Sample.Names.}

\item{Input.DataFrame}{A data frame imported from the .xlsx file specifying details of each sample.}

\item{Config.List}{A list containing configuration parameters for tasAnalyzer.}
}
\value{
A list of filtered sequences, in class ShortRead.
}
\description{
Filters merged paired end read files. Supports 5' or 3' indexing barcodes or unique molecular identifiers (UMIs).
}
