% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{tas_filter_count}
\alias{tas_filter_count}
\title{Counts number of filtered reads at each step of analysis.}
\usage{
tas_filter_count(
  Sample.Names,
  Reads.Filtered.List,
  Sequence.Table.List,
  Input.DataFrame,
  Config.List,
  WD
)
}
\arguments{
\item{Sample.Names}{A character vector with names of all samples.}

\item{Reads.Filtered.List}{A list of filtered reads in ShortReads format (output from tas_filter). Names of each list entry must correspond to Sample.Names.}

\item{Sequence.Table.List}{A list of data frames containing each unique DNA sequence for all samples. Names of each list entry must correspond to Sample.Names.}

\item{Input.DataFrame}{The data frame imported from the .xlsx file specifying details of each sample.}

\item{Config.List}{A list containing configuration parameters for tasAnalyzer.}

\item{WD}{A character vector indicating the file path to the working directory containing input files.}
}
\value{
A data frame with read filtration counts.
}
\description{
After filtration and sequence processing, counts the number of filtered reads after each step. Writes the output to logs/Filtering.xlsx.
}
