% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutations.R
\name{tas_find_AID_targets}
\alias{tas_find_AID_targets}
\alias{tas_find_AID_Targets}
\title{Finds and annotates AID hotspot motifs.}
\usage{
tas_find_AID_Targets(Sample.Names, Reference.Sequences.DNA, Config.List)
}
\arguments{
\item{Sample.Names}{A character vector with names of all samples.}

\item{Reference.Sequences.DNA}{A list containing the expected DNA sequences for each sample.}

\item{Config.List}{A list containing configuration parameters for tasAnalyzer.}
}
\value{
A list of AID hotspot coordinates for each sample.
}
\description{
Uses the reference sequences of each sample being analyzed to identify AID hotspot motifs. Identification is based on both the common WRCY and the slightly more inclusive WRCH motif. Evaluates both strands.
}
