% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{tas_import}
\alias{tas_import}
\title{Read input .xlsx file.}
\usage{
tas_import(Input.File, config)
}
\arguments{
\item{Input.File}{Complete file path of the import .csv file.}

\item{config}{A character vector provided by parent funtion tas_analyze. Can be 'sheet', 'manual', or 'shiny'.}
}
\value{
A data frame containing all input parameters and configurations required for analysis.
}
\description{
Reads the input .csv file and imports all parameters that will be used for downstream analysis.
}
