% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandaseq.R
\name{tas_import_fastq}
\alias{tas_import_fastq}
\title{Reads .fastq files of merged paired end reads.}
\usage{
tas_import_fastq(Sample.Names, Fastq.File.Path)
}
\arguments{
\item{Sample.Names}{A character vector with names of all samples.}

\item{Fastq.File.Path}{A character vector with file paths for each of the merged .fastq.gz files. Names must be identical to Sample.Names.}
}
\value{
A list of merged reads for each sample, in ShortRead format.
}
\description{
Reads the output .fastq files containing paired end reads created by tas_pandaseq_terminal for all samples. Imports as ShortRead format.
}
