% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutations.R
\name{tas_measure_mutations}
\alias{tas_measure_mutations}
\title{Measure mutations in sequence tables}
\usage{
tas_measure_mutations(
  Sample.Names,
  Sequence.Table.List,
  AID.Targets,
  Input.DataFrame,
  Config.List
)
}
\arguments{
\item{Sample.Names}{A character vector with names of all samples.}

\item{Sequence.Table.List}{A list of data frames containing each unique DNA sequence for all samples. Names of each list entry must correspond to Sample.Names.}

\item{AID.Targets}{A list of AID hotspot coordinates for each sample.}

\item{Input.DataFrame}{A data frame imported from the .xlsx file specifying details of each sample.}

\item{Config.List}{A list containing configuration parameters for tasAnalyzer.}
}
\value{
A list providing the analyzed outputs for each sample.
}
\description{
Uses provided reference DNA sequences and the sequence tables generated during filtration to measure mutations in the sequence population. Optional modules support analysis of AID hotspot motifs, breakdown by antibody regions if applicable, and the translated amino acid sequence.
}
