% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msa.R
\name{tas_msa}
\alias{tas_msa}
\title{Perform multiple sequence alignments of filtered and annotated sequences}
\usage{
tas_msa(Sample.Names, Sequence.Table.List, Input.DataFrame, Config.List)
}
\arguments{
\item{Sample.Names}{A character vector with names of all samples.}

\item{Sequence.Table.List}{A list of data frames containing each unique DNA sequence for all samples. Names of each list entry must correspond to Sample.Names.}

\item{Input.DataFrame}{A data frame imported from the .xlsx file specifying details of each sample.}

\item{Config.List}{A list containing configuration parameters for tasAnalyzer.}
}
\value{
A list containing ggplot objects of the specified analyses for each sample.
}
\description{
Uses labeled sequence tables to perform sequence alignments of top common sequences. The number of sequences to be aligned is specified in Config.List by sequence.alignment.count. Can be resource intensive, and increases greatly with more sequences. A default of 10 sequences is recommended as reasonable to run locally and also for readability of output figures. Will automatically trim sequence numbers if fewer sequence species are included for certain samples. Also supports optional amino acid analysis with the protein.mutations parameter or generation of phylogenetic trees with the PhyloTree parameter in Config.List.
}
