% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyMerge.R
\name{tas_pandaseq_wasm}
\alias{tas_pandaseq_wasm}
\title{Helper function for generation of PANDAseq arguments.}
\usage{
tas_pandaseq_wasm(
  Sample.Names,
  Input.DataFrame,
  Config.List,
  WD,
  shiny.env,
  shiny.fileTable
)
}
\arguments{
\item{Sample.Names}{A character vector with names of all samples.}

\item{Input.DataFrame}{A data frame imported from the .xlsx file specifying details of each sample.}

\item{Config.List}{A list containing configuration parameters for tasAnalyzer.}

\item{WD}{A character vector indicating the file path to the working directory containing input files.}

\item{shiny.env}{A logical (TRUE/FALSE) identifying whether processing is being perfomed by the Shiny app. Assigned by parent functions.}

\item{shiny.fileTable}{A data.table with information about uploaded files in the Shiny app. Assigned automatically by parent functions.}
}
\value{
A list where each element contains arguments to feed into the pandaseq function.
}
\description{
This function generates a list containing a set of arguments for each sample in the input .csv file. Helper function to tas_shiny_args.
}
