% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{tas_seq_table}
\alias{tas_seq_table}
\title{Makes data frames with details of all filtered sequences for each sample.}
\usage{
tas_seq_table(Sample.Names, Reads.Filtered.List, Input.DataFrame, Config.List)
}
\arguments{
\item{Sample.Names}{A character vector with names of all samples.}

\item{Reads.Filtered.List}{A list of filtered reads in ShortReads format (output from tas_filter). Names of each list entry must correspond to Sample.Names.}

\item{Input.DataFrame}{A data frame imported from the .xlsx file specifying details of each sample.}

\item{Config.List}{A list containing configuration parameters for tasAnalyzer.}
}
\value{
A list of sequence tables (data frames) for each sample. If UMIs were included, also returns a similar table with all sequences for counting filter statistics.
}
\description{
Processes filtered reads to generate data frames with details of all filtered sequences for each sample. Supports samples with or without unique moleculear identifiers (UMIs). If UMIs are included, sequences will be processed based on UMIs. Otherwise, unique sequences are merged based on read count.
}
