% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyMerge.R
\name{tas_shiny_args}
\alias{tas_shiny_args}
\title{Generate a list pandaseq arguments for the shinylive app.}
\usage{
tas_shiny_args(
  InputFilePath,
  config = "shiny",
  shiny.settings = list(),
  shiny.env = TRUE,
  shiny.fileTable = character(),
  WorkingDirectory = NULL
)
}
\arguments{
\item{InputFilePath}{The complete file path of the input .csv file. Make sure all R1 and R2 files are in the same directory.}

\item{config}{Determines where the function draws the analysis parameters from. 'manual' allows paramaters to be specified in the function, whereas 'shiny' is used by the shiny app.}

\item{shiny.settings}{A list of function settings set by the shiny app.}

\item{shiny.env}{A logical (TRUE/FALSE) to tell whether the function is being executed by the shiny app (TRUE) or not (FALSE).}

\item{shiny.fileTable}{A character vector containing names of files uploaded to the shiny app.}
}
\value{
A list where each element contains arguments to feed into the pandaseq function.
}
\description{
This function generates a list containing a set of arguments for each sample in the input .csv file. This is designed to return to the shinylive app, which uses the list to invoke an Enscripten-compiled fork of PANDAseq through a Javascript intermediary.
}
