% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{tas_write_filtered}
\alias{tas_write_filtered}
\title{Writes and compresses .fastq files after filtering.}
\usage{
tas_write_filtered(Sample.Names, Reads.Filtered.List, Config.List, WD)
}
\arguments{
\item{Sample.Names}{A character vector with names of all samples.}

\item{Reads.Filtered.List}{A list of filtered reads in ShortReads format (output from tas_filter). Names of each list entry must correspond to Sample.Names.}

\item{Config.List}{A list containing configuration parameters for tasAnalyzer.}

\item{WD}{A character vector indicating the file path to the working directory containing input files.}
}
\value{
No returns within R. Writes files to disk.
}
\description{
After sequences are filtered, writes .fastq files of filtered reads, and compresses files as .gz files.
}
