% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmpliconSequencing.R
\name{umi_pileup}
\alias{umi_pileup}
\title{Merge data.table by UMIs}
\usage{
umi_pileup(x)
}
\arguments{
\item{x}{Input vector of sequences to compare.}
}
\value{
Either an optimal consensus sequence for the UMI, or "Rejected".
}
\description{
Helper function for tas_sequence_table. Determines the optimal sequence for each UMI and filters non-passing UMIs as "Rejected".
}
