% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeAmplicon.R
\name{analyzeAmplicon}
\alias{analyzeAmplicon}
\alias{vanalyzeAmplicon}
\title{Analyze NGS amplicons}
\usage{
analyzeAmplicon(input.settings, ...)

vanalyzeAmplicon(settings.list, rows = NULL)
}
\arguments{
\item{input.settings}{A set of tasaR settings. Can be either a directory path to a .csv file (see makeSettingsCSV)}

\item{...}{additional arguments passed through to sub-functions}

\item{settings.list}{A list of settings}

\item{rows}{Optional vector of row numbers to analyze in a .csv input}
}
\value{
An S4 object of class AmpliconSequencing

A list of AmpliconSequencing analysis objects
}
\description{
Takes a set of tasaR settings and performs a complete analysis pipeline. \emph{input.settings} can be a tas.object.settings object, a compliant R list or data.frame, or a file.path to a .csv file.
See also \code{\link{buildSettings}}, \code{\link{readSettings}}, and \code{\link{makeSettingsCSV}}.
}
\section{Functions}{
\itemize{
\item \code{vanalyzeAmplicon()}: Analyze multiple amplicons in sequence

}}
\section{Additional arguments}{

\describe{
\item{\code{row = NULL}}{set the row number if reading settings from a .csv file}
}
}

\section{settings.list formats}{

\describe{
\item{\verb{A list of tas.object.settings objects}}{}
\item{\verb{A file path to a .csv file and a vector of row numbers}}{}
\item{\verb{A list of settings as R lists}}{}
\item{\verb{A data.frame of settings, each row is a sample}}{}
}
}

\examples{
analyzeAmplicon('/.../folder/settings.csv', row = 2)
analyzeAmplicon(data.frame)
analyzeAmplicon(list)
vanalyzeAmplicon(list(tas.object.settings))
vanalyzeAmplicon('/.../folder/settings.csv', rows = 2:11)
vanalyzeAmplicon(list(list))
#' vanalyzeAmplicon(data.frame)
}
