% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeAmplicon.R
\name{analyzeAmplicon}
\alias{analyzeAmplicon}
\title{Analyze NGS amplicons}
\usage{
analyzeAmplicon(input.settings, ...)
}
\arguments{
\item{input.settings}{A set of tasaR settings. Can be either a directory path to a .csv file (see makeSettingsCSV)}

\item{...}{additional arguments passed through to sub-functions}
}
\value{
An S4 object of class AmpliconSequencing
}
\description{
Takes a set of tasaR settings and performs a complete analysis pipeline. \emph{input.settings} can be a tas.object.settings object, a compliant R list or data.frame, or a file.path to a .csv file.
See also \code{\link{buildSettings}}, \code{\link{readSettings}}, and \code{\link{makeSettingsCSV}}.
}
\section{Additional arguments}{

\describe{
\item{\code{row = NULL}}{set the row number if reading settings from a .csv file}
}
}

\examples{
\dontrun{
  analyzeAmplicon('/.../folder/settings.csv', row = 2)
  analyzeAmplicon(data.frame)
  analyzeAmplicon(list)
}
}
