% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-coercion.R
\name{as.list.tas.object.settings}
\alias{as.list.tas.object.settings}
\alias{as.list,tasaR-method}
\title{Coerce tas.object.settings to list}
\usage{
\method{as.list}{tas.object.settings}(x, ...)
}
\arguments{
\item{x}{A tas.object.settings object}
}
\description{
Coerce tas.object.settings to list
}
