% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{as.numeric,tas.dna.repair-method}
\alias{as.numeric,tas.dna.repair-method}
\alias{as.numeric,tasaR-method}
\title{Coerce tas.dna.repair to numeric vector}
\usage{
\S4method{as.numeric}{tas.dna.repair}(x)
}
\arguments{
\item{x}{A tas.dna.repair object}
}
\description{
Coerce tas.dna.repair to numeric vector
}
