% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchAnalyzeAmplicon.R
\name{batchSummarize}
\alias{batchSummarize}
\title{Summary analyses for batch processing}
\usage{
batchSummarize(
  results.list,
  export = FALSE,
  path = NULL,
  suppressConsoleOutput = FALSE
)
}
\arguments{
\item{results.list}{A list of AmpliconSequencing objects from \code{\link{batchAnalyzeAmplicon}}}

\item{export}{A logical specifying whether the resulting tables and graphs should also be written to disk}

\item{path}{(Optional) If \code{export = TRUE}, the file path to save the exported information to}

\item{suppressConsoleOutput}{Logical vector specifying whether to print results to console or not, if applicable}
}
\value{
A list of results, with sub-lists for tables and graphs
}
\description{
Performs comparison analyses of the results from \code{\link{batchAnalyzeAmplicon}}. Results are returned
as an R list, and can be saved as a variable. Also includes the option to directly export the results
to a folder specified by the user. Treats all samples equally, so is unaware of pairings of control/
experimental samples to direct analysis.
}
\examples{
\dontrun{
  batchSummarize(results.list)
  batchSummarize(results.list, export = TRUE, path = './save_in_folder/')
}
}
