% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildSequenceTable.R
\name{buildSequenceTable}
\alias{buildSequenceTable}
\title{Make a labeled table of sequences}
\usage{
buildSequenceTable(settings, ..., progress_callback = NULL)
}
\arguments{
\item{settings}{An object of S4 class tas.settings}

\item{...}{Additional paramaters to be passed to sub-function \code{sequenceTable}. Also inherited from parent functions.}

\item{progress_callback}{Optional function receiving structured progress events during sequence table construction}
}
\value{
An S4 object of class tas.sequences
}
\description{
Uses the settings specified by a tas.settings object, including a file path to a merged .fastq file, to generate a labeled table of sequences, with counts, frequencies, DNA mutations, and AA mutations annotated.

Will use barcodes for sequence filtering/demultiplexing if provided. Will bin sequences and count by UMIs if provided (only supports 1 UMI currently). Otherwise, will bin sequences and filter based on the global setting tasGlobalSettings$read.frequency.limit (the minimum \% to accept a sequence).
}
\examples{
\dontrun{
  buildSequenceTable(settings)
}
}
