% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSettings.R
\name{buildSettings}
\alias{buildSettings}
\title{Make a settings object for tasaR}
\usage{
buildSettings(
  Name = "unknown",
  IsAntibody = FALSE,
  MergedFASTQPath,
  ReferenceSequence,
  ForwardPrimer,
  ReversePrimer,
  AmpliconLength,
  InsertStart,
  InsertEnd,
  ForwardExtensionType = "",
  ForwardExtension = "",
  ReverseExtensionType = "",
  ReverseExtension = "",
  FR1Start = NA_integer_,
  CDR1Start = NA_integer_,
  FR2Start = NA_integer_,
  CDR2Start = NA_integer_,
  FR3Start = NA_integer_,
  CDR3Start = NA_integer_,
  FR4Start = NA_integer_,
  FR4End = NA_integer_
)
}
\arguments{
\item{Name}{Character vector of sample name}

\item{IsAntibody}{Logical vector of if sequence is an antibody}

\item{MergedFASTQPath}{Character vector of file path to the .fastq file to analyze}

\item{ReferenceSequence}{Character vector of the expected amplicon sequence}

\item{ForwardPrimer}{Character vector of the forward primer DNA sequence that \emph{binds to the amplicon}}

\item{ReversePrimer}{Character vector of the reverse primer DNA sequence that \emph{binds to the amplicon}}

\item{AmpliconLength}{Total length of the amplicon (exluding sequencing adapters)}

\item{InsertStart}{Integer of the first nt of the reference sequence}

\item{InsertEnd}{Integer of the last nt of the reference sequence. Should be a negative number, count \emph{backwards} from the 3' end.}

\item{ForwardExtensionType}{Character vector of the type of extension on the forward primer. Can be ("", "Barcode", or "UMI").}

\item{ForwardExtension}{Character vector of the DNA sequence of the forward extension. Accepts standard ambiguity codes.}

\item{ReverseExtensionType}{Character vector of the type of extension on the forward primer. Can be ("", "Barcode", or "UMI").}

\item{ReverseExtension}{Character vector of the DNA sequence of the reverse extension. Accepts standard ambiguity codes.}

\item{FR1Start}{Integer of first nt of the FR1 region if sequence is an antibody. Defaults to NA.}

\item{CDR1Start}{Integer of first nt of the CDR1 region if sequence is an antibody. Defaults to NA.}

\item{FR2Start}{Integer of first nt of the FR2 region if sequence is an antibody. Defaults to NA.}

\item{CDR2Start}{Integer of first nt of the CDR2 region if sequence is an antibody. Defaults to NA.}

\item{FR3Start}{Integer of first nt of the FR3 region if sequence is an antibody. Defaults to NA.}

\item{CDR3Start}{Integer of first nt of the CDR3 region if sequence is an antibody. Defaults to NA.}

\item{FR4Start}{Integer of first nt of the FR4 region if sequence is an antibody. Defaults to NA.}

\item{FR4End}{Integer of last nt of the FR4 region if sequence is an antibody. Defaults to NA.}
}
\value{
An S4 object of class tas.object.settings
}
\description{
Make a settings object for tasaR
}
\examples{
buildSettings(Name = "test", IsAntibody = TRUE, MergedFASTQPath = file.path(tempdir(), "test-merged.fastq"), ReferenceSequence = "GTTCAACTGGTGGAAAGCGG...",
              ForwardPrimer = "GTAAAACGACGGCCAGT", ReversePrimer = "CAGGAAACAGCTATGAC", AmpliconLength = 395L, InsertStart = 25L, InsertEnd = -30L,
              ForwardExtensionType = "Barcode", ForwardExtension = "GCTAGCC",  ReverseExtensionType = "UMI", ReverseExtension = "NNNYRNNNYRNN",
              FR1Start = 1L, CDR1Start = 73L, FR2Start = 97L, CDR2Start = 148L, FR3Start = 172L, CDR3Start = 286L, FR4Start = 310L, FR4End = 342L)

}
