% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmpliconSequencing.R
\name{defineSettingsObject}
\alias{defineSettingsObject}
\title{Define an S4 object of class tas.object.settings}
\usage{
defineSettingsObject(df, snames)
}
\arguments{
\item{df}{A data.frame}

\item{snames}{A character vector of slot names for tas.object.settings defined by parent function readSettings().}
}
\value{
An S4 object of class tas.object.settings
}
\description{
Helper function for readSettings().
}
