% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics-results.R
\name{exportTables}
\alias{exportTables}
\alias{exportTables,AmpliconSequencing,character-method}
\alias{exportTables,tas.sequences,character-method}
\alias{exportTables,tas.mutations,character-method}
\alias{exportTables,tas.dna.repair,character-method}
\title{Export tables of tasaR analysis}
\usage{
\S4method{exportTables}{AmpliconSequencing,character}(results, path, ...)

\S4method{exportTables}{tas.sequences,character}(results, path, ...)

\S4method{exportTables}{tas.mutations,character}(results, path, ...)

\S4method{exportTables}{tas.dna.repair,character}(results, path, ...)
}
\arguments{
\item{results}{S4 object of class AmpliconSequencing, tas.mutation, tas.sequences, or tas.dna.repair}

\item{path}{Character vector specifying the file path to write the .csv files to.}

\item{...}{Additional parameters}
}
\description{
Exports tables of the results of tasaR analyses to a series of .csv files, and copies them to a path specified by the user.
}
\section{Output Tables}{

\describe{
\item{\code{Sequence Table.csv}}{Overview table showing seqences, counts, mutations, and translations. From tas.mutations.}
\item{\code{Read Counts.csv}}{Table showing the read counts at different stages of filtering.}
\item{\code{All DNA Mutations.csv}}{Table showing the \% DNA mutation at each nt position.}
\item{\code{Cytosine DNA Mutations.csv}}{Table showing the \% DNA mutation at each AID cytosine position.}
\item{\code{Non-Cytosine DNA Mutations.csv}}{Table showing the \% DNA mutation at each nt position that is not an AID cytosine.}
\item{\code{Motif Sums.csv}}{Table summarizing the frequency of mutation at different motifs and denominators.}
\item{\code{All Protein Mutations.csv}}{Table showing the \% protein mutation at each amino acid position.}
\item{\code{Protein Mutation Matrix.csv}}{Table showing the frequency of each amino acid occurring at each position in the protein sequence.}
\item{\code{WRCH Table.csv}}{Table summarizing the position of AID hotspots (WRCH motif) and mutation frequency at the cytosine.}
\item{\code{WRCY Table.csv}}{Table summarizing the position of AID hotspots (WRCY motif) and mutation frequency at the cytosine.}
\item{\code{DNA Repair Types.csv}}{Table showing the inferred frequency of different DNA repair pathways..}
}
}

