% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{getAlignments}
\alias{getAlignments}
\alias{getAlignments,tas.alignment-method}
\alias{getAlignments,AmpliconSequencing-method}
\alias{getDNAalign}
\alias{getDNAalign,tas.alignment-method}
\alias{getDNAalign,AmpliconSequencing-method}
\alias{getAAalign}
\alias{getAAalign,tas.alignment-method}
\alias{getAAalign,AmpliconSequencing-method}
\title{Retrieve pairwise alignments}
\usage{
\S4method{getAlignments}{tas.alignment}(object)

\S4method{getAlignments}{AmpliconSequencing}(object)

\S4method{getDNAalign}{tas.alignment}(object)

\S4method{getDNAalign}{AmpliconSequencing}(object)

\S4method{getAAalign}{tas.alignment}(object)

\S4method{getAAalign}{AmpliconSequencing}(object)
}
\arguments{
\item{object}{An S4 object of class tas.alignment or AmpliconSequencing}
}
\description{
Extracts DNA and/or protein (AA) pairwise alignments from an S4 object of class tas.alignment or its parent class AmpliconSequencing.
}
\section{Methods (by class)}{
\itemize{
\item \code{getAlignments(tas.alignment)}: Method for class tas.alignment

\item \code{getAlignments(AmpliconSequencing)}: Method for class AmpliconSequencing

\item \code{getDNAalign(tas.alignment)}: Method for class tas.alignment

\item \code{getDNAalign(AmpliconSequencing)}: Method for class AmpliconSequencing

\item \code{getAAalign(tas.alignment)}: Method for class tas.alignment

\item \code{getAAalign(AmpliconSequencing)}: Method for class AmpliconSequencing

}}
\section{Functions}{
\itemize{
\item \code{getAlignments()}: Extracts a list of DNA and AA pairwise alignments

\item \code{getDNAalign()}: Extracts a pairwise alignment of DNA sequences

\item \code{getAAalign()}: Extracts a pairwise alignment of protein sequences

}}
\examples{
getAlignments(object)
getDNAalign(object)
getAAalign(object)
}
