% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics-tas.mutations.R
\docType{methods}
\name{getMutations}
\alias{getMutations}
\alias{getMutations,tasaR-method}
\alias{getDNAMutations}
\alias{getDNAMutations,tas.mutations-method}
\alias{getDNAMutations,AmpliconSequencing-method}
\alias{getMutationDistributionDNA}
\alias{getMutationDistributionDNA,tas.mutations-method}
\alias{getMutationDistributionDNA,AmpliconSequencing-method}
\alias{getMutationDistributionCytosine}
\alias{getMutationDistributionCytosine,tas.mutations-method}
\alias{getMutationDistributionCytosine,AmpliconSequencing-method}
\alias{getMutationDistributionNonCytosine}
\alias{getMutationDistributionNonCytosine,tas.mutations-method}
\alias{getMutationDistributionNonCytosine,AmpliconSequencing-method}
\alias{getMutationMotifSums}
\alias{getMutationMotifSums,tas.mutations-method}
\alias{getMutationMotifSums,AmpliconSequencing-method}
\alias{getMutationDistributionAA}
\alias{getMutationDistributionAA,tas.mutations-method}
\alias{getMutationDistributionAA,AmpliconSequencing-method}
\alias{getMutationMatrixAA}
\alias{getMutationMatrixAA,tas.mutations-method}
\alias{getMutationMatrixAA,AmpliconSequencing-method}
\alias{getAIDTables}
\alias{getAIDTables,tas.mutations-method}
\alias{getAIDTables,AmpliconSequencing-method}
\alias{getWRCHTable}
\alias{getWRCHTable,tas.mutations-method}
\alias{getWRCHTable,AmpliconSequencing-method}
\alias{getWRCYTable}
\alias{getWRCYTable,tas.mutations-method}
\alias{getWRCYTable,AmpliconSequencing-method}
\alias{getAbMutations}
\alias{getAbMutations,tas.mutations-method}
\alias{getAbMutations,AmpliconSequencing-method}
\title{Mutation retrieval methods}
\usage{
\S4method{getDNAMutations}{tas.mutations}(object)

\S4method{getDNAMutations}{AmpliconSequencing}(object)

\S4method{getMutationDistributionDNA}{tas.mutations}(object)

\S4method{getMutationDistributionDNA}{AmpliconSequencing}(object)

\S4method{getMutationDistributionCytosine}{tas.mutations}(object)

\S4method{getMutationDistributionCytosine}{AmpliconSequencing}(object)

\S4method{getMutationDistributionNonCytosine}{tas.mutations}(object)

\S4method{getMutationDistributionNonCytosine}{AmpliconSequencing}(object)

\S4method{getMutationMotifSums}{tas.mutations}(object)

\S4method{getMutationMotifSums}{AmpliconSequencing}(object)

\S4method{getMutationDistributionAA}{tas.mutations}(object)

\S4method{getMutationDistributionAA}{AmpliconSequencing}(object)

\S4method{getMutationMatrixAA}{tas.mutations}(object)

\S4method{getMutationMatrixAA}{AmpliconSequencing}(object)

\S4method{getAIDTables}{tas.mutations}(object)

\S4method{getAIDTables}{AmpliconSequencing}(object)

\S4method{getWRCHTable}{tas.mutations}(object)

\S4method{getWRCHTable}{AmpliconSequencing}(object)

\S4method{getWRCYTable}{tas.mutations}(object)

\S4method{getWRCYTable}{AmpliconSequencing}(object)

\S4method{getAbMutations}{tas.mutations}(object)

\S4method{getAbMutations}{AmpliconSequencing}(object)
}
\arguments{
\item{object}{An S4 object of class tas.mutations or AmpliconSequencing}
}
\description{
Methods to retrieve mutation information from an object of class tas.mutations or AmpliconSequencing
}
\section{Functions}{
\itemize{
\item \code{getDNAMutations()}: Extracts a list containing DNA mutation information

\item \code{getMutationDistributionDNA()}: Extracts a data.frame quantifying mutation rate at all positions in the sequence

\item \code{getMutationDistributionCytosine()}: Extracts a data.frame quantifying mutation rate at all AID hotspot cytosines in the sequence

\item \code{getMutationDistributionNonCytosine()}: Extracts a data.frame quantifying mutation rate at all bases that are not AID hotspot cytosines

\item \code{getMutationMotifSums()}: Extracts a vector summarizing mutations at different motifs and denominators

\item \code{getMutationDistributionAA()}: Extracts a data.frame of protein mutation positions and frequencies

\item \code{getMutationMatrixAA()}: Extracts a matrix of protein mutation identities and frequencies

\item \code{getAIDTables()}: Extracts a list containing tables for WRCH and WRCY AID hotspot motifs

\item \code{getWRCHTable()}: Extracts a data.frame showing positions and mutations at WRCH AID hotspot motifs

\item \code{getWRCYTable()}: Extracts a data.frame showing positions and mutations at WRCY AID hotspot motifs

\item \code{getAbMutations()}: Extracts antibody-specific mutation information

}}
\examples{
\dontrun{
  getDNAMutations(object)
}
\dontrun{
  getMutationDistributionDNA(object)
}
\dontrun{
  getMutationDistributionCytosine(object)
}
\dontrun{
  getMutationDistributionNonCytosine(object)
}
\dontrun{
  getMutationMotifSums(object)
}
\dontrun{
  getMutationDistributionAA(object)
}
\dontrun{
  getMutationMatrixAA(object)
}
\dontrun{
  getAIDTables(object)
}
\dontrun{
  getWRCHTable(object)
}
\dontrun{
  getWRCYTable(object)
}
\dontrun{
  getAbMutations(object)
}
}
