% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmpliconSequencing.R
\name{getMutations}
\alias{getMutations}
\alias{getMutations,tas.mutations-method}
\alias{getMutations,AmpliconSequencing-method}
\alias{getMutationDistribution}
\alias{getMutationDistribution,tas.mutations-method}
\alias{getMutationDistribution,AmpliconSequencing-method}
\alias{getMutationDistributionCytosine}
\alias{getMutationDistributionCytosine,tas.mutations-method}
\alias{getMutationDistributionCytosine,AmpliconSequencing-method}
\alias{getMutationDistributionNonCytosine}
\alias{getMutationDistributionNonCytosine,tas.mutations-method}
\alias{getMutationDistributionNonCytosine,AmpliconSequencing-method}
\alias{getMutationMotifSums}
\alias{getMutationMotifSums,tas.mutations-method}
\alias{getMutationMotifSums,AmpliconSequencing-method}
\title{Retrieve mutation distributions}
\usage{
\S4method{getMutations}{tas.mutations}(object)

\S4method{getMutations}{AmpliconSequencing}(object)

\S4method{getMutationDistribution}{tas.mutations}(object)

\S4method{getMutationDistribution}{AmpliconSequencing}(object)

\S4method{getMutationDistributionCytosine}{tas.mutations}(object)

\S4method{getMutationDistributionCytosine}{AmpliconSequencing}(object)

\S4method{getMutationDistributionNonCytosine}{tas.mutations}(object)

\S4method{getMutationDistributionNonCytosine}{AmpliconSequencing}(object)

\S4method{getMutationMotifSums}{tas.mutations}(object)

\S4method{getMutationMotifSums}{AmpliconSequencing}(object)
}
\arguments{
\item{object}{An S4 object of class tas.mutations or AmpliconSequencing}
}
\description{
Extracts quantifications of mutations from an S4 object of class tas.mutations or its parent class AmpliconSequencing.
}
\section{Methods (by class)}{
\itemize{
\item \code{getMutations(tas.mutations)}: Method for class tas.mutations

\item \code{getMutations(AmpliconSequencing)}: Method for class AmpliconSequencing

\item \code{getMutationDistribution(tas.mutations)}: Method for class tas.mutations

\item \code{getMutationDistribution(AmpliconSequencing)}: Method for class AmpliconSequencing

\item \code{getMutationDistributionCytosine(tas.mutations)}: Method for class tas.mutations

\item \code{getMutationDistributionCytosine(AmpliconSequencing)}: Method for class AmpliconSequencing

\item \code{getMutationDistributionNonCytosine(tas.mutations)}: Method for class tas.mutations

\item \code{getMutationDistributionNonCytosine(AmpliconSequencing)}: Method for class AmpliconSequencing

\item \code{getMutationMotifSums(tas.mutations)}: Method for class tas.mutations

\item \code{getMutationMotifSums(AmpliconSequencing)}: Method for class AmpliconSequencing

}}
\section{Functions}{
\itemize{
\item \code{getMutations()}: Extracts a list containing mutation information

\item \code{getMutationDistribution()}: Extracts a data.frame quantifying mutation rate at all positions in the sequence

\item \code{getMutationDistributionCytosine()}: Extracts a data.frame quantifying mutation rate at all AID hotspot cytosines in the sequence

\item \code{getMutationDistributionNonCytosine()}: Extracts a data.frame quantifying mutation rate at all bases that are not AID hotspot cytosines

\item \code{getMutationMotifSums()}: Extracts a vector summarizing mutations at different motifs and denominators

}}
\examples{
getMutations(object)
getMutationDistribution(object)
getMutationDistributionCytosine(object)
getMutationDistributionNonCytosine(object)
getMutationMotifSums(object)
}
