% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics-tas.mutations.R
\docType{methods}
\name{getMutations}
\alias{getMutations}
\alias{getMutations,tasaR-method}
\alias{getDNAMutations}
\alias{getMutationDistributionDNA}
\alias{getMutationDistributionCytosine}
\alias{getMutationDistributionNonCytosine}
\alias{getMutationMotifSums}
\alias{getMutationDistributionAA}
\alias{getMutationMatrixAA}
\alias{getAIDTables}
\alias{getWRCHTable}
\alias{getWRCYTable}
\alias{getAbMutations}
\title{Mutation retrieval methods}
\arguments{
\item{object}{An S4 object of class tas.mutations or AmpliconSequencing}
}
\description{
Methods to retrieve mutation information from an object of class tas.mutations or AmpliconSequencing
}
\section{Functions}{
\itemize{
\item \code{getDNAMutations()}: Extracts a list containing DNA mutation information

\item \code{getMutationDistributionDNA()}: Extracts a data.frame quantifying mutation rate at all positions in the sequence

\item \code{getMutationDistributionCytosine()}: Extracts a data.frame quantifying mutation rate at all AID hotspot cytosines in the sequence

\item \code{getMutationDistributionNonCytosine()}: Extracts a data.frame quantifying mutation rate at all bases that are not AID hotspot cytosines

\item \code{getMutationMotifSums()}: Extracts a vector summarizing mutations at different motifs and denominators

\item \code{getMutationDistributionAA()}: Extracts a data.frame of protein mutation positions and frequencies

\item \code{getMutationMatrixAA()}: Extracts a matrix of protein mutation identities and frequencies

\item \code{getAIDTables()}: Extracts a list containing tables for WRCH and WRCY AID hotspot motifs

\item \code{getWRCHTable()}: Extracts a data.frame showing positions and mutations at WRCH AID hotspot motifs

\item \code{getWRCYTable()}: Extracts a data.frame showing positions and mutations at WRCY AID hotspot motifs

\item \code{getAbMutations()}: Extracts antibody-specific mutation information

}}
\examples{
getDNAMutations(object)
getMutationDistributionDNA(object)
getMutationDistributionCytosine(object)
getMutationDistributionNonCytosine(object)
getMutationMotifSums(object)
getMutationDistributionAA(object)
getMutationMatrixAA(object)
getAIDTables(object)
getWRCHTable(object)
getWRCYTable(object)
getAbMutations(object)
}
