% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{getMutationTypes}
\alias{getMutationTypes}
\alias{getMutationTypes,tas.dna.repair-method}
\alias{getMutationTypes,AmpliconSequencing-method}
\title{Retrieve DNA repair pathway inferences}
\usage{
\S4method{getMutationTypes}{tas.dna.repair}(object)

\S4method{getMutationTypes}{AmpliconSequencing}(object)
}
\arguments{
\item{object}{S4 object of class tas.dna.repair or AmpliconSequencing}
}
\description{
Extracts a table (data.frame) of DNA repair pathway frequencies from an S4 object of class tas.dna.repair or its parent class AmpliconSequencing.
}
\section{Methods (by class)}{
\itemize{
\item \code{getMutationTypes(tas.dna.repair)}: Method for class tas.dna.repair

\item \code{getMutationTypes(AmpliconSequencing)}: Method for class AmpliconSequencing

}}
\examples{
getMutationTypes(object)
}
