% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmpliconSequencing.R
\name{getSequenceTable}
\alias{getSequenceTable}
\alias{getSequenceTable,tas.sequences-method}
\alias{getSequenceTable,AmpliconSequencing-method}
\title{Retrieve table of sequences}
\usage{
\S4method{getSequenceTable}{tas.sequences}(object)

\S4method{getSequenceTable}{AmpliconSequencing}(object)
}
\arguments{
\item{object}{S4 object of class tas.sequences or AmpliconSequencing}
}
\description{
Extracts a table (data.frame) of sequences, including sequence quantification and annotations of DNA and protein mutations detected from an S4 object of class tas.sequences or its parent class AmpliconSequencing.
}
\section{Methods (by class)}{
\itemize{
\item \code{getSequenceTable(tas.sequences)}: Method for class tas.sequences

\item \code{getSequenceTable(AmpliconSequencing)}: Method for class AmpliconSequencing

}}
\examples{
getSequenceTable(object)
}
