% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics-tas.sequences.R
\name{getSequenceTable}
\alias{getSequenceTable}
\alias{getSequenceTable,tas.sequences-method}
\alias{getSequenceTable,AmpliconSequencing-method}
\alias{getSequencesDNA}
\alias{getSequencesDNA,tas.sequences-method}
\alias{getSequencesDNA,AmpliconSequencing-method}
\alias{getSequencesAA}
\alias{getSequencesAA,tas.sequences-method}
\alias{getSequencesAA,AmpliconSequencing-method}
\alias{getSequenceSupplemental}
\alias{getSequenceSupplemental,tas.sequences-method}
\alias{getSequenceSupplemental,AmpliconSequencing-method}
\title{Retrieve sequence information}
\usage{
\S4method{getSequenceTable}{tas.sequences}(object)

\S4method{getSequenceTable}{AmpliconSequencing}(object)

\S4method{getSequencesDNA}{tas.sequences}(object)

\S4method{getSequencesDNA}{AmpliconSequencing}(object)

\S4method{getSequencesAA}{tas.sequences}(object)

\S4method{getSequencesAA}{AmpliconSequencing}(object)

\S4method{getSequenceSupplemental}{tas.sequences}(object)

\S4method{getSequenceSupplemental}{AmpliconSequencing}(object)
}
\arguments{
\item{object}{S4 object of class tas.sequences or AmpliconSequencing}
}
\description{
Extracts a table (data.frame) of sequences, including sequence quantification and annotations of DNA and protein mutations detected from an S4 object of class tas.sequences or its parent class AmpliconSequencing.
}
\section{Methods (by class)}{
\itemize{
\item \code{getSequenceTable(tas.sequences)}: Method for class tas.sequences

\item \code{getSequenceTable(AmpliconSequencing)}: Method for class AmpliconSequencing

\item \code{getSequencesDNA(tas.sequences)}: Method for class tas.sequences

\item \code{getSequencesDNA(AmpliconSequencing)}: Method for class AmpliconSequencing

\item \code{getSequencesAA(tas.sequences)}: Method for class tas.sequences

\item \code{getSequencesAA(AmpliconSequencing)}: Method for class AmpliconSequencing

\item \code{getSequenceSupplemental(tas.sequences)}: Method for class tas.sequences

\item \code{getSequenceSupplemental(AmpliconSequencing)}: Method for class AmpliconSequencing

}}
\section{Functions}{
\itemize{
\item \code{getSequencesDNA()}: Extracts a character vector of unique DNA sequences

\item \code{getSequencesAA()}: Extracts a character vector of unique protein sequences

\item \code{getSequenceSupplemental()}: Extracts a data.table with additional sequence information and metadata

}}
\examples{
getSequenceTable(object)
getSequencesDNA(object)
getSequencesAA(object)
getSequenceSupplemental(object)
}
