% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics-tas.sequences.R
\name{getSequenceTableSimplified}
\alias{getSequenceTableSimplified}
\alias{getSequenceTableSimplified,tas.sequences-method}
\alias{getSequenceTableSimplified,AmpliconSequencing-method}
\alias{getSequenceTableSimplified,PairedAmpliconSequencing-method}
\title{Retrieve a simplified table of sequences}
\usage{
\S4method{getSequenceTableSimplified}{tas.sequences}(object)

\S4method{getSequenceTableSimplified}{AmpliconSequencing}(object)

\S4method{getSequenceTableSimplified}{PairedAmpliconSequencing}(object)
}
\arguments{
\item{object}{S4 object of class tas.sequences, AmpliconSequencing, or PairedAmpliconSequencing}
}
\description{
Retrieve a simplified table of sequences
}
\section{Methods (by class)}{
\itemize{
\item \code{getSequenceTableSimplified(tas.sequences)}: Method for class tas.sequences

\item \code{getSequenceTableSimplified(AmpliconSequencing)}: Method for class AmpliconSequencing

\item \code{getSequenceTableSimplified(PairedAmpliconSequencing)}: Method for class PairedAmpliconSequencing

}}
\section{Functions}{
\itemize{
\item \code{getSequenceTableSimplified()}: Extracts a sequence table and simplifies entries for a heterozygous reference sequence.
For instance, a control heterozygous sample might have 2 distinct wild-type (WT) alleles. This function would combine those
into a single entry and add their proportions, giving the total frequency of each mutation regardless of the underlying allele.
Sequence information is removed because it differs between paired alleles.

}}
\examples{
getSequenceTableSimplified(object)
}
