% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmpliconSequencing.R
\name{getSettings}
\alias{getSettings}
\alias{getSettings,AmpliconSequencing-method}
\title{Retrieve object settings}
\usage{
\S4method{getSettings}{AmpliconSequencing}(object)

\S4method{getSettings}{AmpliconSequencing}(object)
}
\arguments{
\item{object}{S4 object of class tas.object.settings or AmpliconSequencing}
}
\description{
Extracts a list of tasAnalyzer settings from an S4 object of class tas.object.settings or its parent class AmpliconSequencing.
}
\section{Methods (by class)}{
\itemize{
\item \code{getSettings(AmpliconSequencing)}: Method for class AmpliconSequencing

\item \code{getSettings(AmpliconSequencing)}: Method for class AmpliconSequencing

}}
\examples{
getSettings(object)
}
