% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics-tas.object.settings.R
\name{getSettings}
\alias{getSettings}
\alias{getSettings,tas.object.settings-method}
\alias{getSettings,AmpliconSequencing-method}
\title{Retrieve object settings}
\usage{
\S4method{getSettings}{tas.object.settings}(object)

\S4method{getSettings}{AmpliconSequencing}(object)
}
\arguments{
\item{object}{S4 object of class tas.object.settings or AmpliconSequencing}
}
\description{
Extracts a list of tasaR settings from an S4 object of class tas.object.settings or its parent class AmpliconSequencing.
}
\section{Methods (by class)}{
\itemize{
\item \code{getSettings(tas.object.settings)}: Method for class tas.object.settings

\item \code{getSettings(AmpliconSequencing)}: Method for class AmpliconSequencing

}}
\examples{
\dontrun{
  getSettings(object)
}
}
