% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{graphResults}
\alias{graphResults}
\title{Graph results of tasaR analysis}
\arguments{
\item{results}{S4 object of class AmpliconSequencing, tas.mutation, tas.sequences, or tas.dna.repair}

\item{output}{Character vector specifying the type of graph to make}

\item{labelCDRs}{(Optional) Logical vector specifying whether the CDR regions (if an antibody) should be labeled on the resulting graphs}

\item{mutationTypes}{(Optional) Character vector specifying the type of DNA mutation summary graph to make}
}
\value{
A ggplot2 graph as specified in the arguments
}
\description{
Graphs a variety of data outputs from tasaR analysis. Graph type is set based on
the 'output' parameter.
}
\section{Output Types}{

\describe{
\item{\code{dna.positions}}{A bar chart showing the frequency of mutations at each nucleotide position.}
\item{\code{dna.positions.labeled}}{A bar chart showing the frequency of mutations at each nucleotide position. AID cytosines are highlighted in red.}
\item{\code{aa.positions}}{A bar chart showing the frequency of mutations at each amino acid position.}
\item{\code{aa.positions.labeled}}{A bar chart showing the frequency of mutations at each amino acid position. Amino acids whose codon includes an AID cytosine are highlighted in red.}
\item{\code{aid.boxplot}}{A box plot comparing the mutation frequency at AID cytosines to all other bases.}
\item{\code{aa.mutations}}{A bar chart showing the frequency and identity of all amino acid mutations at each amino acid position.}
\item{\code{mutation.types.stacked}}{Charts showing the frequency of inferred DNA repair types. Using the mutationTypes argument, can generate a "bar" (stacked bar graph), "pie" chart, or "donut" chart.}
\item{\code{histogram}}{Histogram showing the count of sequences with differing numbers of mutations.}
\item{\code{dna.align}}{Sequence alignment of top DNA sequences against the reference sequence. Default 10 sequences, change with optional settings seq.number = x.}
\item{\code{aa.align}}{Sequence alignment of top protein sequences against the reference sequence. Default 10 sequences, change with optional settings seq.number = x.}
}
}

\section{Input data types for results}{

\describe{
\item{\code{AmpliconSequencing}}{A complete AmpliconSequencing object contains the data to generate each type of graph.}
\item{\code{tas.mutations}}{A complete tas.mutations object contains the data to generate all graphs except for stacked mutation types and histogram. If labeling CDRs or AID cytosines for position graphs, a settings object of tas.object.settings or an appropriately formatted list or data.frame must also be provided.}
\item{\code{tas.sequences}}{Contains the data to generate a histogram.}
\item{\code{tas.dna.repair}}{Contains the data to generate mutation type graphs.}
}
}

