% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{isEmpty-tasaR}
\alias{isEmpty-tasaR}
\alias{isEmpty,tasaR-method}
\alias{isEmpty,tas.object.settings-method}
\alias{isEmpty,tas.mutations-method}
\alias{isEmpty,tas.sequences-method}
\alias{isEmpty,tas.dna.repair-method}
\alias{isEmpty,AmpliconSequencing-method}
\title{isEmpty methods for package tasaR}
\usage{
\S4method{isEmpty}{tas.object.settings}(x)

\S4method{isEmpty}{tas.mutations}(x)

\S4method{isEmpty}{tas.sequences}(x)

\S4method{isEmpty}{tas.dna.repair}(x)

\S4method{isEmpty}{AmpliconSequencing}(x)
}
\description{
Extends the isEmpty function to S4 objects in package tasaR to detect whether S4 classes in tasa are empty, holding only initialized values.
}
