% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmpliconSequencing.R
\name{makeSettingsCSV}
\alias{makeSettingsCSV}
\title{Write a .CSV file to input settings}
\usage{
makeSettingsCSV(filedest)
}
\arguments{
\item{filedest}{Destination for the .csv file to be written to. May be a folder or file path ("./" or "./X.csv"). If folder, will be automatically named 'settings.csv'.}
}
\description{
Write a .CSV file to input settings
}
\examples{
makeSettingsCSV('/.../folder')
makeSettingsCSV('/.../folder/file.csv')
}
