% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measureDNArepair.R
\name{measureDNArepair}
\alias{measureDNArepair}
\title{Estimate DNA repair pathway usage}
\usage{
measureDNArepair(sequence.table, settings)
}
\arguments{
\item{sequence.table}{S4 object of class tas.sequences}

\item{settings}{S4 object of class tas.object.settings}
}
\value{
S4 object of class tas.dna.repair with estimates of DNA repair types
}
\description{
Estimates the frequency of DNA repair pathway usage in a given sequence table of DNA sequences. Classification as follows:

\emph{WT}: unmodified wild-type reference sequences

\emph{NHEJ}: non-homologous end joining, insertions and deletions of -1 and -2

\emph{MMEJ}: microhomology-mediated end joining, deletions larger than -2

\emph{BaseChange}: base mutations without an indel scar

\emph{IndelBaseChange}: sequences with both an indel and base changes

\emph{Other}: other sequences (containing multiple indels, etc.)
}
\examples{
\dontrun{
  measureDNArepair(sequence.table, settings)
}
}
