% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pandaseq.R
\name{pandaseq_merge_files}
\alias{pandaseq_merge_files}
\title{Merge paired end reads}
\usage{
pandaseq_merge_files(
  forward_fastq,
  reverse_fastq,
  output_fastq,
  log_file = NULL,
  min_length = NULL,
  max_length = NULL,
  extra_args = c("-F", "-d", "bFSrk"),
  verbose = FALSE
)
}
\arguments{
\item{forward_fastq}{Character vector, file path to the forward (R1) read.}

\item{reverse_fastq}{Character vector, file path to the reverse (R2) read.}

\item{output_fastq}{Character vector, file path to write the merged read to. Sub-directories will be automatically created.}

\item{log_file}{Character vector, file path for the log file to be written to. Sub-directories will be automatically created.}

\item{min_length}{Numeric vector, minimum sequence length to allow.}

\item{max_length}{Numeric vector, maximum sequence length to allow.}

\item{extra_args}{Character vector, additional arguments are set with reasonable defaults. See PANDAseq manual for more details.}

\item{verbose}{Logical vector, prints verbose output resulst from PANDAseq if TRUE.}
}
\description{
A function to merge paired end reads using an included installation of the software PANDAseq.
Writes a merged fastq file to disk with the given arguments.
}
