% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmpliconSequencing.R
\docType{class}
\name{tas.alignment-class}
\alias{tas.alignment-class}
\title{S4 class tas.alignment}
\description{
S4 class containing pairwise alignments of sequences against the reference. Contains both DNA and protein (AA) alignments.
}
\section{Slots}{

\describe{
\item{\code{DNA}}{S4 object of class PairwiseAlignmentsSingleSubject for DNA sequences}

\item{\code{AA}}{S4 object of class PairwiseAlignmentsSingleSubject for protein sequences}
}}

