% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-tas.dna.repair.R
\docType{class}
\name{tas.dna.repair-class}
\alias{tas.dna.repair-class}
\title{S4 class tas.dna.repair}
\description{
S4 object showing the inferred frequency of different types of DNA repair
}
\section{Slots}{

\describe{
\item{\code{WT}}{numeric vector of percent of unmutated wild-type (WT) sequences}

\item{\code{NHEJ}}{numeric vector of percent of non-homologous end joining (NHEJ) repair (insertion, -1, or -2)}

\item{\code{MMEJ}}{numeric vector of percent of microhomology-mediated end joining (MMEJ) repair (< -2)}

\item{\code{BaseChange}}{numeric vector of percent of sequences with mutated bases but no indels}

\item{\code{IndelBaseChange}}{numeric vector of percent of sequences with mutated bases and indels}

\item{\code{Other}}{numeric vector of percent of sequences not fitting into any other category above}
}}

