% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-tas.mutations.R
\docType{class}
\name{tas.mutations-class}
\alias{tas.mutations-class}
\title{S4 class tas.mutations}
\description{
S4 object of sequence mutations
}
\section{Slots}{

\describe{
\item{\code{DNA}}{a list of DNA mutation information}

\item{\code{AA}}{a list of protein mutation information}

\item{\code{AIDTables}}{a list of summary information of mutations at AID hotspots}

\item{\code{Antibody}}{a list of antibody-specific mutation information}
}}

\section{DNA}{

\describe{
\item{\code{AllMutations}}{data.frame showing position and frequency of mutations at all positions}
\item{\code{CytosineMutations}}{data.frame showing position and frequency of mutations at AID cytosines}
\item{\code{NonCytosineMutations}}{data.frame showing position and frequency of mutations at all positions that are not AID cytosines}
\item{\code{MotifSums}}{numeric vector summarizing the frequency of mutations among different motifs and denominators}
}
}

\section{AA}{

\describe{
\item{\code{AllMutations}}{data.frame showing position and frequency of mutations at all positions}
\item{\code{MutationMatrix}}{matrix showing frequency and identity of all protein mutations (WT omitted)}
}
}

\section{AIDTables Columns}{

\describe{
\item{\code{Motif}}{character vector DNA motif (top strand = WRCH/WRCY, bottom strand = DGYW/RGYW)}
\item{\code{Start}}{numeric vector of starting coordinates of AID motifs}
\item{\code{End}}{numeric vector of ending coordinates of AID motifs}
\item{\code{Cytosine}}{numeric vector of coordinates of AID cytosines}
\item{\code{CytosineMutationFrequency}}{numeric vector of the mutation frequency at the AID cytosine}
}
}

\section{Antibody}{

\describe{
\item{\code{RegionMutations}}{numeric vector of percent deviation from expected mutation rate for each antibody region, assuming an equal distribution across length}
}
}

