% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{tas.mutations-class}
\alias{tas.mutations-class}
\title{S4 class tas.mutations}
\description{
S4 object of sequence mutations
}
\section{Slots}{

\describe{
\item{\code{DNA}}{a list of DNA mutation information}

\item{\code{AA}}{a list of protein mutation information}

\item{\code{AIDTables}}{a list of summary information of mutations at AID hotspots}
}}

\section{DNA}{

\describe{
\item{AllMutations}{data.frame showing position and frequency of mutations at all positions}
\item{CytosineMutations}{data.frame showing position and frequency of mutations at AID cytosines}
\item{NonCytosineMutations}{data.frame showing position and frequency of mutations at all positions that are not AID cytosines}
\item{MotifSums}{numeric vector summarizing the frequency of mutations among different motifs and denominators}
}
}

\section{AA}{

\describe{
\item{AllMutations}{data.frame showing position and frequency of mutations at all positions}
\item{MutationMatrix}{matrix showing frequency and identity of all protein mutations (WT omitted)}
}
}

\section{AIDTables Columns}{

\describe{
\item{Motif}{character vector DNA motif (top strand = WRCH/WRCY, bottom strand = DGYW/RGYW)}
\item{Start}{numeric vector of starting coordinates of AID motifs}
\item{End}{numeric vector of ending coordinates of AID motifs}
\item{Cytosine}{numeric vector of coordinates of AID cytosines}
\item{CytosineMutationFrequency}{numeric vector of the mutation frequency at the AID cytosine}
}
}

