% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{tas.sequences-class}
\alias{tas.sequences-class}
\title{S4 class tas.sequences}
\description{
S4 object of analyzed sequence table
}
\section{Slots}{

\describe{
\item{\code{Table}}{A data.frame with a variable number of rows and 7 named columns:
\describe{
\item{\code{Sequences}}{character vector of DNA sequences}
\item{\code{Count}}{numeric vector of sequence counts}
\item{\code{Percent}}{numeric vector of sequence frequencies}
\item{\code{Indels}}{character vector annotating insertions and deletions in the sequence}
\item{\code{BasesChanged}}{integer vector counting the number of mutated DNA bases in the sequence}
\item{\code{AA}}{character vector of protein sequences}
\item{\code{ProteinMutation}}{character vector annotating protein mutations in the sequence}
}}

\item{\code{Alignments}}{A list of DNA and protein (AA) pairwise alignments}

\item{\code{Supplemental}}{A data.table of DNA sequences paired with UMIs and Illumina sequence IDs}
}}

