% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{tas_analyze}
\alias{tas_analyze}
\title{One-step analysis and export of targeted amplicon sequencing experiments}
\usage{
tas_analyze(
  InputFilePath,
  config = "manual",
  WorkingDirectory = NULL,
  merge.reads = TRUE,
  measure.shm = TRUE,
  sequence.alignment.count = 10,
  read.frequency.limit = 0.001,
  protein.mutations = TRUE,
  PhyloTree = TRUE,
  dna.repair.pathways = TRUE
)
}
\arguments{
\item{InputFilePath}{The complete file path of the input .csv file. Make sure all R1 and R2 files are in the same directory.}

\item{config}{Determines where the function draws the analysis parameters from. 'manual' allows paramaters to be specified in the function.}

\item{WorkingDirectory}{Optional. The path to the folder containing the input .xlsx file and all input .fastq files.}

\item{merge.reads}{A logical (TRUE/FALSE) to tell whether PANDAseq should be used to merge paired end read files.}

\item{measure.shm}{A logical (TRUE/FALSE) to tell whether the somatic hypermutation analysis module should be run.}

\item{sequence.alignment.count}{Enter a whole number value to determine how many sequences to display as multiple sequence alignments. Default value of 10 is recommended, can require significant resources for larger alignments.}

\item{read.frequency.limit}{If UMIs are not used, sequences will be filtered using this number as the minimum read frequency percentage for a sequence to be included in the analysis. Values of 0.1-0.001 are recommended.}

\item{protein.mutations}{A logical (TRUE/FALSE) to tell whether mutations in the amino acid sequences should be analyzed.}

\item{PhyloTree}{A logical (TRUE/FALSE) to tell whether phylogenetic trees should be made for the number of sequences specified by sequence.alignment.count.}

\item{dna.repair.pathways}{A logical (TRUE/FALSE) to tell whether this is a gene editing experiment and the type of mutations observed should be classified and counted by the likely underlying DNA repair pathway.}
}
\value{
A list containing tables and graphical analyses generated by the software.
}
\description{
Full analysis of a targeted amplicon sequencing experiment on Illumina, starting from R1 and R2 .fastq reads.
}
