% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{tas_check}
\alias{tas_check}
\title{Check for presence and formatting of all input files.}
\usage{
tas_check(InputFilePath, Input.DataFrame, Config.List, WD)
}
\arguments{
\item{InputFilePath}{The complete file path of the input .csv file. Make sure all R1 and R2 files are in the same directory.}

\item{Input.DataFrame}{A data frame imported from the .csv file specifying details of each sample.}

\item{Config.List}{A list containing configuration parameters for tasAnalyzer.}

\item{WD}{A character vector indicating the file path to the working directory containing input files.}
}
\value{
No returns within R.
}
\description{
Checks whether InputDataFrame and Config.List contain all expected columns/entries. Also checks whether all unpaired read files specified have been provided.
}
