% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{tas_label_table}
\alias{tas_label_table}
\title{Labels sequence tables with mutations.}
\usage{
tas_label_table(
  Sample.Names,
  Sequence.Table.List,
  Reference.Sequences.DNA,
  Input.DataFrame,
  Config.List
)
}
\arguments{
\item{Sample.Names}{A character vector with names of all samples.}

\item{Sequence.Table.List}{A list of data frames containing each unique DNA sequence for all samples. Names of each list entry must correspond to Sample.Names.}

\item{Reference.Sequences.DNA}{A list of DNAString objects containing expected DNA sequences for all samples. Names of each list entry must correspond to Sample.Names.}

\item{Input.DataFrame}{A data frame imported from the .xlsx file specifying details of each sample.}

\item{Config.List}{A list containing configuration parameters for tasAnaly <- er.}
}
\value{
A list of labeled sequence tables (data frames) for each sample.
}
\description{
Classifies each unique sequence for each sample based on indel type and number of mutated bases. If the amino acid (AA) module is enabled, also determines protein sequences and mutations for each DNA sequence.
}
