% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandaseq.R
\name{tas_pandaseq_terminal}
\alias{tas_pandaseq_terminal}
\title{Runs PANDAseq to merge paired end reads on systems with a Unix terminal (Linux and MacOS).}
\usage{
tas_pandaseq_terminal(
  Sample.Names,
  Input.DataFrame,
  Config.List,
  WD,
  shiny.env,
  shiny.fileTable
)
}
\arguments{
\item{Sample.Names}{A character vector with names of all samples.}

\item{Input.DataFrame}{A data frame imported from the .xlsx file specifying details of each sample.}

\item{Config.List}{A list containing configuration parameters for tasAnalyzer.}

\item{WD}{A character vector indicating the file path to the working directory containing input files.}

\item{shiny.env}{A logical (TRUE/FALSE) identifying whether processing is being perfomed by the Shiny app. Assigned by parent functions.}

\item{shiny.fileTable}{A data.table with information about uploaded files in the Shiny app. Assigned automatically by parent functions.}
}
\value{
No returns within R. Writes files to disk.
}
\description{
Uses information in the input .xlsx file to construct arguments for PANDAseq for each sample as specified. Writes commands to terminal for execution. Compresses the output .fastq files as .gz. Only supported on Linux and MacOS systems with a Unix terminal.
}
